+++ aliding-20230426/ChangeLog.md
@@ -1,3 +1,13 @@
+2024-10-23 Version: 2.26.1
+- Update API QueryGroupLiveInfo: update param AccountContext.
+- Update API QueryGroupLiveInfo: update param AnchorUnionId.
+- Update API QueryGroupLiveInfo: update param AnchorUnionId.
+- Update API QueryGroupLiveInfo: update param LiveUuid.
+- Update API QueryGroupLiveInfo: update param LiveUuid.
+- Update API QueryGroupLiveInfo: update param TenantContext.
+- Update API QueryGroupLiveInfo: update response param.
+
+
 2024-10-18 Version: 2.26.0
 - Support API AddTicketMemo.
 - Support API AssignTicket.